#137 setting up UI for draft on the session overview
