+++ client/src/SessionView/Overview/Drafting.js
@@ -1,0 +1,3 @@
+export function Drafting({ editable, session }) {
+  return <p>hello world drafting</p>
+}
