+++ server/Domain/GameEvents/GameStarted.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
@@ -32,7 +33,7 @@ namespace server.Domain
         {
             get
             {
-                return SetupType == "draft";
+                return !Factions.Any();
             }
         }
     }
